pyproject.toml
[tool.poetry]
name = "idpmodem"
version = "1.1.0"
description = "A library for interfacing with an Inmarsat IsatData Pro satellite modem using serial AT commands."
authors = ["G.Bruce-Payne <geoff.bruce-payne@inmarsat.com>"]
license = "Apache-2.0"
readme = "README.rst"
repository = "https://github.com/Inmarsat/idpmodem"
documentation = "https://inmarsat.github.io/idpmodem"
keywords = ["satellite", "iot", "idp", "isatdata pro"]

[tool.poetry.dependencies]
python = "^3.7"
aioserial = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pylint = "^2.6.0"
virtualenvwrapper = "^4.8.4"
Sphinx = "^3.4.1"
pytest-asyncio = "^0.14.0"
pytest-mock = "^3.5.0"
asynctest = "^0.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
